Float to doubles

+++ UndeadInfection/src/Model/GridMPI.h
@@ -44,7 +44,7 @@ private:
 	void ownershipExchange(int myID, MPI_Datatype agentDatatype, Agent** gridB, MPI_Status& status, int* localStats);
 	void applyBoundary(Agent** gridB, int* localStats);
 	void resolveHumanZombie(Agent &human, Agent &zombie, RandomClass& random, int *localStats);
-	void resolveGridHumanHuman(Agent &agentA,int i, int j, Agent** gridA, RandomClass& random, int* localStats, float probHumanHaveBaby);
+	void resolveGridHumanHuman(Agent &agentA,int i, int j, Agent** gridA, RandomClass& random, int* localStats, double probHumanHaveBaby);
 	void printState(int tick, int* stats);
 
 public:
